Update dep (#15)
* Update probe-rs dependency

* Cleanup

* Ran cargo-fmt

* Add note about sub command alias in README

* Change version number to `0.2.2`

+++ src/commands/debug_response.rs
@@ -42,7 +42,7 @@ pub enum DebugResponse {
     ClearBreakpoint,
     ClearAllBreakpoints,
     Code {
-        pc: u32,
+        pc: u64,
         instructions: Vec<(u32, String)>,
     },
     Stack {
